Structuring the project using clean code principles
